requirements.txt
imapclient==3.0.1
requests==2.32.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
pymysql==1.1.2
